Close #528: Templatize the model number type
The fundamental type for all the physics params used to be hardcoded as
float. Now it can be changed (with a #define), and it's double by default.

Surprisingly, the overall FPS hasn't changed *at all* after this change,
even on my oldish, slow-CPU TOSHIBA laptop! :-o

(NOTE: snapshot save/load can only handle the old format with floats yet!
See #561 for that!)

+ Minor random corollary changes here and there.

+++ src/Engine/SimApp.hpp
@@ -189,14 +189,14 @@ public:
 		return false;
 	}
 	*/
-	virtual bool collide_hook(Model::World* w, Entity* obj1, Entity* obj2, float distance);
+	virtual bool collide_hook(Model::World* w, Entity* obj1, Entity* obj2, double distance);
 	virtual bool touch_hook(Model::World* w, Entity* obj1, Entity* obj2);
 
 	// High-level, abstract (not as in "generic", but "app-level") hook for n-body interactions:
 	// `event` represents the physical property/condition that made it think these might interact.
 	//!!NOTE: This will change to the objects themselves being notified (not the game "superclass")!
-	virtual void undirected_interaction_hook(Model::World* w, Entity* obj1, Entity* obj2, float dt, float distance, ...);
-	virtual void directed_interaction_hook(Model::World* w, Entity* source, Entity* target, float dt, float distance, ...);
+	virtual void undirected_interaction_hook(Model::World* w, Entity* obj1, Entity* obj2, float dt, double distance, ...);
+	virtual void directed_interaction_hook(Model::World* w, Entity* source, Entity* target, float dt, double distance, ...);
 
 
 	//----------------------------------------------------------------------------
